client/game-painter: Tune the maybe_click parameters
When testing with a low-DPI tablet device, the finger position moves
quite a lot when you try to click with your thumb. This was making it a
bit difficult to click on the tile bag when you hold the tablet in your
hands and click with your thumb. To make this a bit easier, the click
distance has been moved back up to 3mm. To avoid the other problem of it
being really difficult to move a tile just a little bit, if you now hold
the finger down for more than a second it will no longer consider it a
click.