client/worker: Don’t update poll_fd array in use by poll
The worker was previously storing an array of poll_fds that is written
to in order to update the events for the connection fd whilst
simultaneously being polled on in another thread. This seems to work on
Linux but on Darwin/iOS it seems to always overwrite the array when poll
returns so the updates to the events for the connection fd get
overwritten. In order to avoid this problem the worker now stores the
poll_fds on the stack instead and recreates it before every poll.