client/game-state: Don’t remove conversation ID after setting connection
Previously when the game state is reset for a conversation ID it would
only set the ID on the vsx_connection and then it would clear it from
the game state. The ID would then be later redownloaded from the server
in order to be visible in the invite URL. Keeping it will make it easier
to store the conversation ID in the instance state in a later patch.