app: Use only one of either instance state or invite intent
If the app is opened via an invite URL and then restarted such as by
rotating the screen, the activity will receive both the intent and the
instance state. If the app has moved on since handling the invite such
as by leaving the game then handling this invite intent would
erroneously cause it try to rejoin the old game. In order to try and
detect this situation the Android app now stores the invite URL in the
instance state. If it is the same as the one coming in from the intent
then we can assume the instance state is more up-to-date and ignore the
intent.