apps: The conversation ID has priority over the instance state
Even though both the apps try to avoid loading both the instance state
and the invite URL, this can still happen if the new invite URL is set
after loading the activity or game view controller but before the game
state is created. In this case the invite URL is more recent than the
instance state so we want to give priority to that.