tile-tool: Split the texture into two
Instead of having one texture with all the letters from Shavian and the
Latin alphabet, there is now one for each. The game is very unlikely to
need both alphabets in the same frame so it’s probably more efficient to
use separate textures. This also avoids increasing the maximum texture
dimensions used by the app.